Refactor starfield component into modular classes
